Move methods to BL from DAO
https://github.com/metasfresh/metasfresh/issues/6485
