+++ de.metas.business/src/main/java/de/metas/pricing/service/IPriceListBL.java
@@ -1,12 +1,5 @@
 package de.metas.pricing.service;
 
-import java.time.ZonedDateTime;
-
-import javax.annotation.Nullable;
-
-import org.compiere.model.I_M_PriceList;
-import org.compiere.model.I_M_PriceList_Version;
-
 import de.metas.currency.CurrencyPrecision;
 import de.metas.lang.SOTrx;
 import de.metas.location.CountryId;
@@ -14,6 +7,12 @@ import de.metas.pricing.PriceListId;
 import de.metas.pricing.PricingSystemId;
 import de.metas.util.ISingletonService;
 import lombok.NonNull;
+import org.compiere.model.I_M_PriceList;
+import org.compiere.model.I_M_PriceList_Version;
+
+import javax.annotation.Nullable;
+import java.time.LocalDate;
+import java.time.ZonedDateTime;
 
 /**
  * @author RC
@@ -42,10 +41,27 @@ public interface IPriceListBL extends ISingletonService
 	 * @param processedPLVFiltering if not <code>null</code>, then only PLVs which have the give value in their <code>Processed</code> column are considered.
 	 *                              task 09533: the user doesn't know about PLV's processed flag, so in most cases we can't filter by it
 	 */
-	@Nullable I_M_PriceList_Version getCurrentPriceListVersionOrNull(
+	@Nullable
+	I_M_PriceList_Version getCurrentPriceListVersionOrNull(
 			@Nullable PricingSystemId pricingSystemId,
 			CountryId countryId,
 			ZonedDateTime date,
 			@Nullable SOTrx soTrx,
 			Boolean processedPLVFiltering);
+
+	/**
+	 * Please keep in sync with {@link #updateAllPLVName(String, PriceListId)} )}
+	 */
+	String createPLVName(final @NonNull String priceListName, @NonNull LocalDate date);
+
+	/**
+	 * Update the Name of all Price List Versions.
+	 * <p>
+	 * The name has the format "Price List.Name + PLV.ValidFrom".
+	 * <p>
+	 * Please keep in sync with {@link #createPLVName(String, LocalDate)}
+	 *
+	 * @return number of updated records
+	 */
+	int updateAllPLVName(final String namePrefix, final PriceListId priceListId);
 }
